fixup! Fix: abort on unaligned pointer in worker threads on Win

+++ src/node_contextify.cc
@@ -260,7 +260,7 @@ BaseObjectPtr<ContextifyContext> ContextifyContext::New(
                               sandbox_obj);
   if (node_is_nwjs) {
     void* data = env->context()->GetAlignedPointerFromEmbedderData(2); //v8ContextPerContextDataIndex
-    if (!((int)data & 1))
+    if (!((uintptr_t)data & 1))
       v8_context->SetAlignedPointerInEmbedderData(2, data);
     v8_context->SetAlignedPointerInEmbedderData(50, (void*)0x08110800);
   }
